Waste Segregation - Copy/requirements.txt
Flask==3.0.3
Pillow==10.4.0
numpy==2.1.2
scikit-learn==1.5.2
joblib==1.4.2
